JBRes-6959: Introduce Add-Comment Transformation (#25)
* feat: add `java` submodule

* feat: create `JavaTransformation` interface in `java` submodule

* feat: create `TextBasedTransformation` abstract class

* feat: create `TransformationExecutor` interface

* feat: create `IntelliJAwareTransformation` interface

* feat: implement `IntelliJTransformationExecutor`

* feat: add `HeadlessLogger` that mirrors logs into stdout/stderr

* feat: implement `AddCommentTransformation` to serve as example

* feat: apply given transformations in `TransformationService`

* feat: use `AddCommentTransformation` in `HeadlessModeStarter`

* feat: remove `parseConfig` from `Transformation`

Because this parsing can be done in init block of an exact transformation.
This makes the API easier since a user doesn't need to explicitly parse the config
(it's responsibility of the transformation itself).

* feat: register transformations in `HeadlessModeStarter`

Additionally, register `AddCommentTransformation` as an example.

* feat: add `add-comment-transformation` in yaml config

* fix: remove dependency on `core` in `java` module

Otherwise, the build fails due to missing credentials in `java` module
required to install a private Space module. It is so because `java` module
used `core` as `implementation` dependency, all `core`'s deps must be resolvable, which isn't the case for `grazie-llm-interaction`.

A possible solution is to duplicate private deps setup from `core`'s build.gradle.kts into `java`'s build file. Probably, there's a better one.
Right now, merely removing a dependency on `core` in `java` module setup.

* feat: remove `name` member field of `Transformation`